Move UTF8Control to util package
