+++ client/src/main/java/pl/korotkevics/ships/client/gui/controllers/MainController.java
@@ -9,6 +9,7 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.Pane;
 import lombok.Getter;
 import pl.korotkevics.ships.client.client.Client;
+import pl.korotkevics.ships.client.gui.util.UTF8Control;
 import pl.korotkevics.ships.client.localization.OurLocale;
 import pl.korotkevics.ships.shared.infra.logging.api.Target;
 import pl.korotkevics.ships.shared.infra.logging.core.SharedLogger;
