+++ client/src/main/java/pl/korotkevics/ships/client/gui/util/UTF8Control.java
@@ -1,4 +1,4 @@
-package pl.korotkevics.ships.client.gui.controllers;
+package pl.korotkevics.ships.client.gui.util;
 
 import java.io.IOException;
 import java.io.InputStream;
